refactor: split selection operator types
