+++ src/CADShared/Cad/Editor/Selection/Filters/OpList.cs
@@ -102,59 +102,3 @@ public abstract class OpList : OpLogi
         return Lst.GetEnumerator();
     }
 }
-
-/// <summary>
-/// 逻辑与类
-/// </summary>
-public class OpAnd : OpList
-{
-    /// <summary>
-    /// 符号名
-    /// </summary>
-    public override string Name => "And";
-
-    /// <summary>
-    /// 添加过滤条件
-    /// </summary>
-    /// <param name="value">过滤器对象</param>
-    public override void Add(OpFilter value)
-    {
-        if (value is OpAnd opand)
-        {
-            foreach (var item in opand)
-                Lst.Add(item);
-        }
-        else
-        {
-            Lst.Add(value);
-        }
-    }
-}
-
-/// <summary>
-/// 逻辑或类
-/// </summary>
-public class OpOr : OpList
-{
-    /// <summary>
-    /// 符号名
-    /// </summary>
-    public override string Name => "Or";
-
-    /// <summary>
-    /// 添加过滤条件
-    /// </summary>
-    /// <param name="value">过滤器对象</param>
-    public override void Add(OpFilter value)
-    {
-        if (value is OpOr opor)
-        {
-            foreach (var item in opor)
-                Lst.Add(item);
-        }
-        else
-        {
-            Lst.Add(value);
-        }
-    }
-}
